requirements.txt
discord.py
requests
spotipy
python-dotenv